refactor: move draggable folder support
